apply PEP8 and Google for clang-format

+++ BCPNN/backend/full_cuda_backend/__init__.py
@@ -3,8 +3,16 @@ import numpy as np
 import _bcpnn_backend_full_cuda_internals as backend_cuda
 backend_cuda.initialize()
 
+
 class DenseLayer:
-    def __init__(self, in_features, hypercolumns, minicolumns, taupdt, initial_counters, dtype=None):
+    def __init__(
+            self,
+            in_features,
+            hypercolumns,
+            minicolumns,
+            taupdt,
+            initial_counters,
+            dtype=None):
         self.in_features = in_features
         self.hypercolumns = hypercolumns
         self.minicolumns = minicolumns
@@ -14,8 +22,21 @@ class DenseLayer:
 
         self.initial_counters = initial_counters
 
+
 class StructuralPlasticityLayer:
-    def __init__(self, in_features, hypercolumns, minicolumns, taupdt, khalf, pmin, taubdt, density, mask_iterations, initial_counters, dtype=None):
+    def __init__(
+            self,
+            in_features,
+            hypercolumns,
+            minicolumns,
+            taupdt,
+            khalf,
+            pmin,
+            taubdt,
+            density,
+            mask_iterations,
+            initial_counters,
+            dtype=None):
         self.in_features = in_features
         self.hypercolumns = hypercolumns
         self.minicolumns = minicolumns
@@ -30,6 +51,7 @@ class StructuralPlasticityLayer:
         self.mask_iterations = mask_iterations
 
         self.initial_counters = initial_counters
+
 
 class Network:
     def __init__(self, dtype):
@@ -48,15 +70,37 @@ class Network:
         self._layers.append(layer)
         if isinstance(layer, DenseLayer):
             cs = layer.initial_counters
-            self.net.add_dense_layer(layer.in_features, layer.hypercolumns, layer.minicolumns, layer.taupdt, cs[0], cs[1], cs[2])
+            self.net.add_dense_layer(
+                layer.in_features,
+                layer.hypercolumns,
+                layer.minicolumns,
+                layer.taupdt,
+                cs[0],
+                cs[1],
+                cs[2])
         elif isinstance(layer, StructuralPlasticityLayer):
             cs = layer.initial_counters
-            self.net.add_plastic_layer(layer.in_features, layer.hypercolumns, layer.minicolumns, layer.taupdt, layer.pmin, layer.khalf, layer.taubdt, cs[0], cs[1], cs[2])
+            self.net.add_plastic_layer(
+                layer.in_features,
+                layer.hypercolumns,
+                layer.minicolumns,
+                layer.taupdt,
+                layer.pmin,
+                layer.khalf,
+                layer.taubdt,
+                cs[0],
+                cs[1],
+                cs[2])
         else:
             raise Exception("Unknown layer type:", layer)
 
-    def fit(self, training_data, training_labels, maximal_batch_size, schedule):
-        training_data   = training_data.astype(self.dtype)
+    def fit(
+            self,
+            training_data,
+            training_labels,
+            maximal_batch_size,
+            schedule):
+        training_data = training_data.astype(self.dtype)
         training_labels = training_labels.astype(self.dtype)
         self.net.initiate_training(training_data, training_labels)
 
@@ -67,4 +111,5 @@ class Network:
         images = images.astype(self.dtype)
         labels = labels.astype(self.dtype)
 
-        return self.net.evaluate(images, labels, batch_size) * images.shape[0], images.shape[0]
+        return self.net.evaluate(
+            images, labels, batch_size) * images.shape[0], images.shape[0]
